web_gateway: carve peer_requests.rs (pure move)
Peer/federation HTTP request glue moves to
web_gateway/peer_requests.rs: add/remove-peer payloads + manual peer
persistence, identity summary, send-message/delegate/resolve payloads,
federated answer rewriting, and capability query parsing, with 8
tests. The four request structs' fields widen to pub(crate) for
routes_peers; http_request and setup_peer_op_test stay as shared
fixtures.
